+++ src/Domain/Entities/MediaMetaData/LibraryMediaItemDTO.cs
@@ -284,7 +284,7 @@ public record LibraryMediaItemStreamDTO
     /// <summary>
     /// Unique stream identifier.
     /// </summary>
-    public required long Id { get; init; }
+    public required int Id { get; init; }
 
     /// <summary>
     /// Stream type (1=video, 2=audio, 3=subtitle).
